+++ reductstore/src/replication/remote_bucket/initial_state.rs
@@ -11,7 +11,6 @@ use crate::storage::bucket::RecordRx;
 use async_trait::async_trait;
 use log::error;
 use reduct_base::Labels;
-use reduct_rs::ReductClient;
 use url::Url;
 
 /// Initial state of the remote bucket.
@@ -41,15 +40,18 @@ impl RemoteBucketState for InitialState {
         content_length: u64,
         rx: RecordRx,
     ) -> Box<dyn RemoteBucketState + Sync + Send> {
+        // Try to get the bucket.
         let bucket = self.client.get_bucket(&self.bucket_name).await;
         match bucket {
             Ok(bucket) => {
+                // Bucket is available, transition to the available state and write the record.
                 let new_state = Box::new(BucketAvailableState::new(self.client, bucket));
                 new_state
                     .write_record(entry, timestamp, labels, content_type, content_length, rx)
                     .await
             }
             Err(err) => {
+                // Bucket is unavailable, transition to the unavailable state.
                 error!(
                     "Failed to get remote bucket {}/{}: {}",
                     self.client.url(),
@@ -69,6 +71,9 @@ impl RemoteBucketState for InitialState {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::replication::remote_bucket::tests::{
+        bucket, client, MockReductBucketApi, MockReductClientApi,
+    };
     use rstest::rstest;
 
     #[rstest]
@@ -80,5 +85,29 @@ mod tests {
         assert_eq!(state.ok(), false);
     }
 
-    // TODO: Add more tests with mocks
+    #[rstest]
+    #[tokio::test]
+    async fn test_bucket_available(
+        mut client: MockReductClientApi,
+        mut bucket: MockReductBucketApi,
+    ) {
+        bucket
+            .expect_write_record()
+            .return_once(|_, _, _, _, _, _| Ok(()));
+        client
+            .expect_get_bucket()
+            .return_once(move |_| Ok(Box::new(bucket)));
+
+        let state = InitialState {
+            client: Box::new(client),
+            bucket_name: "test_bucket".to_string(),
+        };
+
+        let (_tx, rx) = tokio::sync::mpsc::channel(1);
+        let state = state
+            .write_record("test_entry", 0, Labels::new(), "text/plain", 0, rx)
+            .await;
+
+        assert_eq!(state.ok(), true);
+    }
 }
